Allow the use of msUPN to be used to specify principals
The msUPN extension (szOID_NT_PRINCIPAL_NAME, 1.3.6.1.4.1.311.20.2.3)
can be used to specify the principal associated with a KCA
certificate.  If this is found in a certificate, assume we are dealing
with a KCA certificate.

In addition, keep track of the issuers and the realms associated
with them so that when we see a certificate from a particular
issuer, we can determine whether that certificate was issued by a
KCA.

Change-Id: I7e0e9244ed5dbf0c1ebc6d7144ca8cee46bd97f1